requirements.txt
aiofiles==23.2.1
aiogram==3.10.0
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.2
annotated-types==0.7.0
attrs==23.2.0
Babel==2.15.0
cachetools==5.4.0
certifi==2024.7.4
environs==11.0.0
fluent-compiler==0.3
fluent.syntax==0.19.0
fluentogram==1.1.7
frozenlist==1.4.1
greenlet==3.0.3
idna==3.7
Jinja2==3.1.4
magic-filter==1.0.12
Mako==1.3.5
MarkupSafe==2.1.5
marshmallow==3.21.3
multidict==6.0.5
ordered-set==4.1.0
packaging==24.1
psycopg==3.2.1
psycopg-binary==3.2.1
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
six==1.16.0
SQLAlchemy==2.0.31
typing_extensions==4.12.2
tzdata==2024.1
watchdog==2.3.1
yarl==1.9.4
